+++ apps/charterafrica/src/components/Comments/Comment.js
@@ -12,114 +12,19 @@ import {
   ListItemAvatar,
   ListItemText,
   SvgIcon,
-  useTheme,
 } from "@mui/material";
 import PropTypes from "prop-types";
-import React, { forwardRef, useState, useRef } from "react";
+import React, { forwardRef, useState } from "react";
+
+import CommentsThread from "./CommentsThread";
 
 import chevronDown from "@/charterafrica/assets/icons/chevron-down, Size=24, Color=CurrentColor.svg";
 import chevronUp from "@/charterafrica/assets/icons/Chevron-up, Size=24, Color=CurrentColor.svg";
 import thumbsUp from "@/charterafrica/assets/icons/Type=thumbs-up, Size=24, Color=CurrentColor.svg";
 import userIcon from "@/charterafrica/assets/icons/Type=user, Size=auto, Color=CurrentColor.svg";
 import { neutral } from "@/charterafrica/colors";
+import DynamicLineClampedTypography from "@/charterafrica/components/Comments/DynamicLineClampedTypography";
 import LineClampedRichTypography from "@/charterafrica/components/LineClampedRichTypography";
-
-const DynamicLineClampedTypography = forwardRef(
-  ({ comment, ...props }, ref) => {
-    const ref1 = useRef();
-    const commentHeight = ref1?.current?.offsetHeight;
-    const currentTheme = useTheme();
-    const lineHeight =
-      currentTheme.typography.p3.lineHeight *
-      currentTheme.typography.p3.fontSize;
-    const showReadMore = commentHeight >= lineHeight * 3;
-
-    const [readMoreOpen, setReadMoreOpen] = useState(false);
-    const lineClamp = !readMoreOpen ? 2 : undefined;
-
-    return (
-      <>
-        <LineClampedRichTypography
-          ref={ref}
-          lineClamp={lineClamp}
-          variant="p3"
-          {...props}
-        >
-          {/* eslint-disable-next-line react/no-danger */}
-          <span dangerouslySetInnerHTML={{ __html: comment }} ref={ref1} />
-        </LineClampedRichTypography>
-
-        {showReadMore ? (
-          <Button
-            onClick={() => setReadMoreOpen((v) => !v)}
-            sx={(theme) => ({
-              ...theme.typography.p2,
-              mt: 2,
-              p: 0,
-              color: theme.palette.neutral.main,
-            })}
-          >
-            {readMoreOpen ? "Read less" : `Read more`}
-          </Button>
-        ) : null}
-      </>
-    );
-  }
-);
-const ThreadChildComment = forwardRef((props, ref) => {
-  const {
-    comment,
-    authorProfileImageUrl,
-    authorDisplayName,
-    publishedAt,
-    updatedAt,
-  } = props;
-
-  return (
-    <Box sx={{ mb: 3 }} ref={ref}>
-      <Box display="flex" alignItems="center">
-        <Avatar
-          sx={(theme) => ({
-            bgcolor: "#fff",
-            color: theme.palette.neutral.main,
-          })}
-          src={authorProfileImageUrl}
-        >
-          <SvgIcon
-            inheritViewBox
-            component={userIcon}
-            sx={{
-              color: neutral[800],
-              fill: "none",
-              height: 40,
-              width: 40,
-            }}
-          />
-        </Avatar>
-        <LineClampedRichTypography
-          lineClamp={1}
-          color="neutral.dark"
-          sx={{ ml: 1.25 }}
-          variant="p2"
-        >
-          {authorDisplayName}
-        </LineClampedRichTypography>
-        <LineClampedRichTypography
-          lineClamp={1}
-          sx={{ ml: 1 }}
-          variant="p2"
-          color="neutral.main"
-        >
-          {publishedAt} {publishedAt !== updatedAt && "(edited)"}
-        </LineClampedRichTypography>
-      </Box>
-
-      <Box sx={{ mt: 1 }}>
-        <DynamicLineClampedTypography comment={comment} />
-      </Box>
-    </Box>
-  );
-});
 
 const Comment = forwardRef((props, ref) => {
   const {
@@ -264,7 +169,7 @@ const Comment = forwardRef((props, ref) => {
                   </AccordionSummary>
                   <AccordionDetails sx={{ p: 0 }}>
                     {threads.map((thread) => (
-                      <ThreadChildComment {...thread} key={thread.id} />
+                      <CommentsThread {...thread} key={thread.id} />
                     ))}
                   </AccordionDetails>
                 </Accordion>
